extract delivery method policy
